backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
fastapi-cors==0.0.6
uagents==0.16.2
google-generativeai==0.8.3
singlestoredb==0.5.1
requests==2.32.3
pydantic==2.9.2
openai==0.26.5
